Updated: * Organization of logic Added: * utils folder
